src/zenml/integrations/gcp/orchestrators/vertex_scheduler/requirements.txt
google-api-python-client>=1.7.8,<2
google-cloud-aiplatform